Split large file
