+++ tests/services/monitor_data_test.py
@@ -14,7 +14,7 @@ import pytest
 
 from ess.livedata.config import instrument_registry, workflow_spec
 from ess.livedata.config.models import ConfigKey
-from ess.livedata.core.job import JobAction, JobCommand
+from ess.livedata.core.job_manager import JobAction, JobCommand
 from ess.livedata.services.monitor_data import make_monitor_service_builder
 from tests.helpers.livedata_app import LivedataApp
